requirements.txt
flask
natsort
flask_profiler
#influxdb
#git+https://github.com/influxdata/influxdb-python.git
influxdb
pymodbus
pyModbusTCP
flasgger
pysnmp==4.4.2
exscript
docopt
pyserial
python-telegram-bot
redis==3.2.0
scrapy
gTTS
requests
html2text
gunicorn
supervisor
flask-login
validators
itsdangerous
celery
transitions
flask_bcrypt
#SQLAlchemy==0.9.10
SQLAlchemy>=1.3.11
#Flask-SQLAlchemy==2.1
Flask-SQLAlchemy
paho-mqtt
redlock
numpy
influxdb
pyyaml
oyaml
pyinstrument
Flask-Migrate
Arrow
nose
requests_mock
